feat: allow scientific price notation
This commit ensures that orchestrators can specify pricing using the
scientific notation.
